requirements.txt
# Web framework and server
fastapi>=0.100.0
uvicorn>=0.23.0

# Templates and configuration
jinja2>=3.1.0
pyyaml>=6.0

# Data validation
pydantic>=2.0.0

# HTTP requests
requests>=2.31.0

# Data visualization and graph attacks
matplotlib>=3.7.0
graph_tiger

# HTML/XML parsing
beautifulsoup4>=4.12.0

# LangChain support (optional)
# Uncomment to enable LangChain integration:
# langchain>=0.1.0
# langchain-openai>=0.1.0
# langchain-community>=0.0.10
lxml>=4.9.0

# PDF processing
pdfminer.six>=20220524
reportlab>=3.6.13

# Frontend UI (optional)
streamlit>=1.39.0

# File upload support
python-multipart>=0.0.9
